+++ src/Extractor/ReadAccessor.php
@@ -188,18 +188,18 @@ final class ReadAccessor
             /*
              * Use the property fetch to read the value
              *
-             * isset($input->property_name)
+             * isset($input->property_name) && null === $input->property_name
              */
-            return new Expr\BooleanNot(new Expr\Isset_([new Expr\PropertyFetch($input, $this->accessor)]));
+            return new Expr\BinaryOp\LogicalAnd(new Expr\BooleanNot(new Expr\Isset_([new Expr\PropertyFetch($input, $this->accessor)])), new Expr\BinaryOp\Identical(new Expr\ConstFetch(new Name('null')), new Expr\PropertyFetch($input, $this->accessor)));
         }
 
         if (self::TYPE_ARRAY_DIMENSION === $this->type) {
             /*
              * Use the array dim fetch to read the value
              *
-             * isset($input['property_name'])
+             * isset($input['property_name']) && null === $input->property_name
              */
-            return new Expr\BooleanNot(new Expr\Isset_([new Expr\ArrayDimFetch($input, new Scalar\String_($this->accessor))]));
+            return new Expr\BinaryOp\LogicalAnd(new Expr\BooleanNot(new Expr\Isset_([new Expr\PropertyFetch($input, $this->accessor)])), new Expr\BinaryOp\Identical(new Expr\ConstFetch(new Name('null')), new Expr\PropertyFetch($input, $this->accessor)));
         }
 
         if (self::TYPE_SOURCE === $this->type) {
@@ -207,6 +207,52 @@ final class ReadAccessor
                 new Expr\ConstFetch(new Name('null')),
                 $input,
             );
+        }
+
+        throw new CompileException('Invalid accessor for read expression');
+    }
+
+    public function getIsUndefinedExpression(Expr\Variable $input): Expr
+    {
+        if (\in_array($this->type, [self::TYPE_METHOD, self::TYPE_SOURCE])) {
+            /*
+             * false
+             */
+            return new Expr\ConstFetch(new Name('false'));
+        }
+
+        if (self::TYPE_PROPERTY === $this->type) {
+            if ($this->private) {
+                /*
+                 * When the property is private we use the extract callback that can read this value
+                 *
+                 * @see \AutoMapper\Extractor\ReadAccessor::getExtractIsUndefinedCallback()
+                 *
+                 * $this->extractIsUndefinedCallbacks['property_name']($input)
+                 */
+                return new Expr\FuncCall(
+                    new Expr\ArrayDimFetch(new Expr\PropertyFetch(new Expr\Variable('this'), 'extractIsUndefinedCallbacks'), new Scalar\String_($this->accessor)),
+                    [
+                        new Arg($input),
+                    ]
+                );
+            }
+
+            /*
+             * Use the property fetch to read the value
+             *
+             * !array_key_exists($property_name, (object) $input)
+             */
+            return new Expr\BooleanNot(new Expr\FuncCall(new Name('array_key_exists'), [new Arg(new Scalar\String_($this->accessor)), new Arg(new Expr\Cast\Array_($input))]));
+        }
+
+        if (self::TYPE_ARRAY_DIMENSION === $this->type) {
+            /*
+             * Use the array dim fetch to read the value
+             *
+             * !array_key_exists('property_name', $input)
+             */
+            return new Expr\BooleanNot(new Expr\FuncCall(new Name('array_key_exists'), [new Arg(new Scalar\String_($this->accessor)), new Arg($input)]));
         }
 
         throw new CompileException('Invalid accessor for read expression');
@@ -256,6 +302,38 @@ final class ReadAccessor
      * Get AST expression for binding closure when dealing with a private property.
      */
     public function getExtractIsNullCallback(string $className): ?Expr
+    {
+        if ($this->type !== self::TYPE_PROPERTY || !$this->private) {
+            return null;
+        }
+
+        /*
+         * Create extract is null callback for this accessor
+         *
+         *  \Closure::bind(function ($object) {
+         *      return !isset($object->property_name) && null === $object->property_name;
+         *  }, null, $className)
+         */
+        return new Expr\StaticCall(new Name\FullyQualified(\Closure::class), 'bind', [
+            new Arg(
+                new Expr\Closure([
+                    'params' => [
+                        new Param(new Expr\Variable('object')),
+                    ],
+                    'stmts' => [
+                        new Stmt\Return_(new Expr\BinaryOp\LogicalAnd(new Expr\BooleanNot(new Expr\Isset_([new Expr\PropertyFetch(new Expr\Variable('object'), $this->accessor)])), new Expr\BinaryOp\Identical(new Expr\ConstFetch(new Name('null')), new Expr\PropertyFetch(new Expr\Variable('object'), $this->accessor)))),
+                    ],
+                ])
+            ),
+            new Arg(new Expr\ConstFetch(new Name('null'))),
+            new Arg(new Scalar\String_($className)),
+        ]);
+    }
+
+    /**
+     * Get AST expression for binding closure when dealing with a private property.
+     */
+    public function getExtractIsUndefinedCallback(string $className): ?Expr
     {
         if ($this->type !== self::TYPE_PROPERTY || !$this->private) {
             return null;